CoinedSolutions-master/requirements.txt
asgiref==3.2.10
boto3==1.16.14
botocore==1.19.14
dj-database-url==0.5.0
Django==3.1.5
django-crispy-forms==1.9.2
django-filter==2.4.0
django-heroku==0.3.1
django-js-asset==1.2.2
django-mptt==0.11.0
django-storages==1.10.1
gunicorn==20.0.4
jmespath==0.10.0
Pillow==8.0.1
psycopg2==2.8.6
python-dateutil==2.8.1
pytz==2020.1
s3transfer==0.3.3
six==1.15.0
sqlparse==0.4.1
urllib3==1.25.11
whitenoise==5.2.0